+++ src/components/spartan/STab/variants/index.ts
@@ -1,0 +1,8 @@
+export { default as UnderlineTab } from './underline/UnderlineTab.vue';
+export { default as UnderlineTabItem } from './underline/UnderlineTabItem.vue';
+
+export { default as PillTab } from './pill/PillTab.vue';
+export { default as PillTabItem } from './pill/PillTabItem.vue';
+
+export { default as VetchTab } from './vetch/VetchTab.vue';
+export { default as VetchTabItem } from './vetch/VetchTabItem.vue';