+++ src/components/spartan/STab/variants/vetch/VetchTab.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts"></script>
+
+<template>
+    <nav aria-label="Tabs">
+        <ul class="flex space-x-2 bg-gray-100 w-fit p-2 border border-gray-300 rounded-xl">
+            <slot />
+        </ul>
+    </nav>
+</template>